+++ .gitignore
@@ -8,4 +8,5 @@ target/
 test/
 .idea
 *.iml
+~$*
 /projects/01-introducao-tdd/nbproject/
