+++ Makefile
@@ -1,0 +1,5 @@
+pdf:
+	asciidoctor-pdf material.adoc
+
+html:
+	asciidoctor -b html material.adoc
